Add index file to manage exports and minor export name updates - Sqon becomes SQON - SQON.has becomes SQON.in - Don't export SQON methods as part of module
